12 delete comment

+++ models/models.js
@@ -112,8 +112,16 @@ const fetchUsers = () => {
 };
 
 const eraseComment = (comment_id) => {
-  return db.query(`DELETE FROM comments WHERE comment_id = $1`, [comment_id]);
+  return db.query(`SELECT * FROM comments WHERE comment_id = $1`, [comment_id])
+    .then((result) => {
+      if (result.rows.length === 0) {
+        return Promise.reject({ status: 404, msg: "Bad request" });
+      } else {
+        return db.query(`DELETE FROM comments WHERE comment_id = $1`, [comment_id]);
+      }
+    });
 }
+
 
 module.exports = {
   fetchCategories,
